refactor: condense sections into reusable traits
